refactor: 하나된 차트들을 각자 기능별로 구분
StackedBarChart,  DoughnutChart를 component로 따로 분리했습니다.
Graph.tsx 내에 해당하는 코드들을 삭제 후 컴포넌트로 대체 했습니다.
